Refactoring: move .cpp and .h files to src/ dir
