--- exception.cpp
@@ -1,1 +1,0 @@
-#include "exception.h"
