+++ src/ganttchart.cpp
@@ -1,0 +1,2 @@
+#include "ganttchart.h"
+
