--- timelineitem.cpp
@@ -1,1 +1,0 @@
-#include "timelineitem.h"
